Created helper class ConnectomeReader for common methods/classes
Consolidated WitvlietDataReader into a single class
