+++ c302/W_SpreadsheetDataReader.py
@@ -1,5 +1,8 @@
-from c302.NeuroMLUtilities import ConnectionInfo
-from c302.NeuroMLUtilities import analyse_connections
+from c302.ConnectomeReader import ConnectionInfo
+from c302.ConnectomeReader import analyse_connections
+from c302.ConnectomeReader import convert_to_preferred_muscle_name
+from c302.ConnectomeReader import is_neuron
+from c302.ConnectomeReader import is_body_wall_muscle
 
 from openpyxl import load_workbook
 import os
@@ -8,214 +11,95 @@ from c302 import print_
 
 spreadsheet_location = os.path.dirname(os.path.abspath(__file__))+"/data/"
 
-def get_old_muscle_name(name):
-    if name.startswith("BWM-VL"):
-        return "MVL%s" %name[6:]
-    elif name.startswith("BWM-VR"):
-        return "MVR%s" %name[6:]
-    elif name.startswith("BWM-DL"):
-        return "MDL%s" %name[6:]
-    elif name.startswith("BWM-DR"):
-        return "MDR%s" %name[6:]
 
-class WitvlietDataReader1:   
+class WitvlietDataReader:   
 
-    def read_data(include_nonconnected_cells=False, neuron_connect=False):
+    def __init__(self, spreadsheet):
 
-        if neuron_connect:
-            conns = []
-            cells = []
-            filename = "%switvliet_2020_7.xlsx"%spreadsheet_location
-            wb = load_workbook(filename)
-            sheet = wb.worksheets[0]
-            print_("Opened the Excel file: " + filename)
+        self.filename = "%s%s"%(spreadsheet_location,spreadsheet)
+
+    def read_data(self,include_nonconnected_cells=True):
+
+        if not include_nonconnected_cells:
+            raise Exception('Option include_nonconnected_cells=False not supported')
             
+        conns = []
+        cells = []
 
-            for row in sheet.iter_rows(min_row=2, values_only=True):  # Assuming data starts from the second row
-                pre = str(row[0])
-                post = str(row[1])
-                syntype = str(row[2])
-                num = int(row[3])
-                synclass = 'Generic_GJ' if 'electrical' in syntype else 'Generic_CS'
-                
+        wb = load_workbook(self.filename)
+        sheet = wb.worksheets[0]
+        print_("Opened the Excel file: " + self.filename)
+        
+        for row in sheet.iter_rows(min_row=2, values_only=True):  # Assuming data starts from the second row
+            #print(row)
+            pre = str(row[0])
+            post = str(row[1])
 
-                conns.append(ConnectionInfo(pre, post, num, syntype, synclass))
-                if pre not in cells:
-                    cells.append(pre)
-                if post not in cells:
-                    cells.append(post)
+            if not is_neuron(pre) or not is_neuron(post):
+                continue  # pre or post is not a neuron
 
-            return cells, conns
+            syntype = str(row[2])
+            num = int(row[3])
+            synclass = 'Generic_GJ' if 'electrical' in syntype else 'Generic_CS'
+            
+            conns.append(ConnectionInfo(pre, post, num, syntype, synclass))
 
-        else:
-            conns = []
-            cells = []
-            filename = "%switvliet_2020_7.xlsx"%spreadsheet_location
-            wb = load_workbook(filename)
-            sheet = wb.worksheets[0]
+            if pre not in cells:
+                cells.append(pre)
+            if post not in cells:
+                cells.append(post)
 
-            print_("Opened Excel file..: " + filename)
-
-            known_nonconnected_cells = ['CANL', 'CANR', 'VC6']
-
-
-            for row in sheet.iter_rows(min_row=2, values_only=True):
-                pre = str(row[0])
-                post = str(row[1])
-                syntype = str(row[2])
-                num = int(row[3])
-                synclass = 'Generic_GJ' if 'electrical' in syntype else 'Generic_CS'
+        return cells, conns
 
 
-                conns.append(ConnectionInfo(pre, post, num, syntype, synclass))
-                if pre not in cells:
-                    cells.append(pre)
-                if post not in cells:
-                    cells.append(post)
-
-            if include_nonconnected_cells:
-                for c in known_nonconnected_cells: cells.append(c)
-
-            return cells, conns
-
-
-    
-    
-    def read_muscle_data():
+    def read_muscle_data(self):
 
         conns = []
         neurons = []
         muscles = []
 
-        filename = "%switvliet_2020_7.xlsx"%spreadsheet_location
-        wb = load_workbook(filename)
+        wb = load_workbook(self.filename)
         sheet = wb.worksheets[0]
 
-        print_("Opened Excel file: "+ filename)
+        print_("Opened Excel file: "+ self.filename)
 
         for row in sheet.iter_rows(min_row=2, values_only=True):
-                pre = str(row[0])
-                post = str(row[1])
-                syntype = str(row[2])
-                num = int(row[3])
-                synclass = 'Generic_GJ' if 'EJ' in syntype else 'Generic_CS'
-                if post.startswith("BWM-"):
-                    post = get_old_muscle_name(post)
-                else:
-                     continue
-                
-                conns.append(ConnectionInfo(pre, post, num, syntype, synclass))
-                if pre not in neurons:
-                        neurons.append(pre)
-                if post not in muscles:
-                        muscles.append(post)
+            pre = str(row[0])
+            post = str(row[1])
+            syntype = str(row[2])
+            num = int(row[3])
+            synclass = 'Generic_GJ' if 'EJ' in syntype else 'Generic_CS'
+
+            if post.startswith("BWM-"):
+                post = convert_to_preferred_muscle_name(post)
+            else:
+                    continue
+            
+            conns.append(ConnectionInfo(pre, post, num, syntype, synclass))
+            if pre not in neurons:
+                    neurons.append(pre)
+            if post not in muscles:
+                    muscles.append(post)
 
 
         return neurons, muscles, conns
     
 def main1():
 
-    cells, neuron_conns = WitvlietDataReader1.read_data(include_nonconnected_cells=True)
-    neurons2muscles, muscles, muscle_conns = WitvlietDataReader1.read_muscle_data()
+    wdr = WitvlietDataReader('witvliet_2020_7.xlsx')
+
+    cells, neuron_conns = wdr.read_data()
+    neurons2muscles, muscles, muscle_conns = wdr.read_muscle_data()
 
     analyse_connections(cells, neuron_conns, neurons2muscles, muscles, muscle_conns)
 
 
-class WitvlietDataReader2:   
-    def read_data(include_nonconnected_cells=False, neuron_connect=False):
-
-        if neuron_connect:
-            conns = []
-            cells = []
-            filename = "%switvliet_2020_8.xlsx"%spreadsheet_location
-            wb = load_workbook(filename)
-            sheet = wb.worksheets[0]
-            print_("Opened the Excel file: " + filename)
-            
-
-            for row in sheet.iter_rows(min_row=2, values_only=True):  # Assuming data starts from the second row
-                pre = str(row[0])
-                post = str(row[1])
-                syntype = str(row[2])
-                num = int(row[3])
-                synclass = 'Generic_GJ' if 'electrical' in syntype else 'Generic_CS'
-
-                conns.append(ConnectionInfo(pre, post, num, syntype, synclass))
-                if pre not in cells:
-                    cells.append(pre)
-                if post not in cells:
-                    cells.append(post)
-
-            return cells, conns
-
-        else:
-            conns = []
-            cells = []
-            filename = "%switvliet_2020_8.xlsx"%spreadsheet_location
-            wb = load_workbook(filename)
-            sheet = wb.worksheets[0]
-
-            print_("Opened Excel file..: " + filename)
-
-            known_nonconnected_cells = ['CANL', 'CANR', 'VC6']
-
-
-            for row in sheet.iter_rows(min_row=2, values_only=True):
-                pre = str(row[0])
-                post = str(row[1])
-                syntype = str(row[2])
-                num = int(row[3])
-                synclass = 'Generic_GJ' if 'electrical' in syntype else 'Generic_CS'
-
-
-                conns.append(ConnectionInfo(pre, post, num, syntype, synclass))
-                if pre not in cells:
-                    cells.append(pre)
-                if post not in cells:
-                    cells.append(post)
-
-            if include_nonconnected_cells:
-                for c in known_nonconnected_cells: cells.append(c)
-
-            return cells, conns
-
-
-    def read_muscle_data():
-
-        conns = []
-        neurons = []
-        muscles = []
-
-        filename = "%switvliet_2020_8.xlsx"%spreadsheet_location
-        wb = load_workbook(filename)
-        sheet = wb.worksheets[0]
-
-        print_("Opened Excel file: "+ filename)
-
-        for row in sheet.iter_rows(min_row=2, values_only=True):
-                pre = str(row[0])
-                post = str(row[1])
-                syntype = str(row[2])
-                num = int(row[3])
-                synclass = 'Generic_GJ' if 'electrical' in syntype else 'Generic_CS'
-                if post.startswith("BWM-"):
-                    post = get_old_muscle_name(post)
-                else:
-                     continue
-
-                conns.append(ConnectionInfo(pre, post, num, syntype, synclass))
-                if pre not in neurons:
-                        neurons.append(pre)
-                if post not in muscles:
-                        muscles.append(post)
-
-
-        return neurons, muscles, conns
-    
 def main2():
 
-    cells, neuron_conns = WitvlietDataReader2.read_data(include_nonconnected_cells=True)
-    neurons2muscles, muscles, muscle_conns = WitvlietDataReader2.read_muscle_data()
+    wdr = WitvlietDataReader('witvliet_2020_8.xlsx')
+
+    cells, neuron_conns = wdr.read_data()
+    neurons2muscles, muscles, muscle_conns = wdr.read_muscle_data()
 
     analyse_connections(cells, neuron_conns, neurons2muscles, muscles, muscle_conns)
